Fix home button in action bar
