+++ AnkiDroid/src/main/java/com/ichi2/anki/Info.java
@@ -164,17 +164,6 @@ public class Info extends AnkiActivity {
         }
     }
 
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            // Respond to the action bar's Up/Home button
-            case android.R.id.home:
-                finishWithAnimation();
-                return true;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-
 
     @Override
     public boolean onKeyDown(int keyCode, KeyEvent event) {
